After extensive checks, the multishift inverter cannot be improved.
changed the test to check what can actually be checked.
If the numerator in the rational approximation is very large, nothing
can be done about it unfortunately. The correction in the inverter for
the convergence criterion tries to cope with that, but for very large
shifts the condition number is so close to 1 that the solver converges
in a few steps, while the estimation of the residual is slightly off.

This is not a problem of the CG-M, it is also a problem of the single
CG.