Fix unit tests
Previously the unit tests used arrays including halos when the code in thie branch expects no halos in the "third" dimension. This produced a bug when the data was given to MPI to perform a halo exchange. This has been fixed so that the size of these arrays is what the rest of the code expects.